Changed 'patients' folder to 'registration' and modified class names for organisation
