+++ utils/constants.js
@@ -1,6 +1,6 @@
-export const API_URL = "https://projectsabai-vza8.onrender.com";
+//export const API_URL = "https://projectsabai-vza8.onrender.com";
 export const CLOUDINARY_URL = "https://res.cloudinary.com/dxy0byphl";
-// export const API_URL = "http://localhost:8000";
+export const API_URL = "http://localhost:8000";
 
 export const venueOptions = {
   SV: "Smong Village",
